+++ CesiumGltfReader/test/TestGltfReader.cpp
@@ -24,6 +24,7 @@
 #include <CesiumUtility/StringHelpers.h>
 
 #include <doctest/doctest.h>
+#include <fmt/format.h>
 #include <glm/ext/matrix_double4x4.hpp>
 #include <glm/ext/vector_float2.hpp>
 #include <glm/ext/vector_float3.hpp>
@@ -473,6 +474,260 @@ TEST_CASE("Read MeshPrimitiveModes") {
       0, 4, 5,
       0, 5, 6,
       0, 6, 1};
+    // clang-format on
+    REQUIRE_EQ(indicesView.size(), int64_t(expected.size()));
+    for (int64_t i = 0; i < indicesView.size(); i++) {
+      REQUIRE_EQ(indicesView[i], expected[size_t(i)]);
+    }
+  }
+}
+
+namespace {
+std::vector<std::byte>
+generateBinaryGltf(int32_t primitiveMode, const std::vector<uint8_t>& indices) {
+  using namespace std::string_literals;
+
+  constexpr std::string_view jsonTemplate = R"(
+    {{
+      "accessors": [
+        {{
+          "bufferView": 0,
+          "count": {},
+          "componentType": 5121,
+          "type": "SCALAR"
+        }}
+      ],
+      "bufferViews": [
+        {{
+          "buffer": 0,
+          "byteLength": {}
+        }}
+      ],
+      "buffers": [
+        {{
+          "byteLength": {}
+        }}
+      ],
+      "meshes": [
+        {{
+          "primitives": [
+            {{
+              "indices": 0,
+              "mode": {}
+            }}
+          ]
+        }}
+      ]
+    }}
+  )";
+
+  size_t indexCount = indices.size();
+  uint32_t indicesByteLength =
+      static_cast<uint32_t>(indexCount * sizeof(uint8_t));
+  std::vector<std::byte> binary(indicesByteLength);
+  std::memcpy(binary.data(), indices.data(), indicesByteLength);
+  if (size_t remainder = binary.size() % 4; remainder > 0) {
+    // Align to 4-byte boundary.
+    binary.resize(binary.size() + (4 - remainder));
+  }
+
+  std::string json = fmt::format(
+      jsonTemplate,
+      indexCount,
+      indicesByteLength,
+      indicesByteLength,
+      primitiveMode);
+  if (size_t remainder = json.size() % 4; remainder > 0) {
+    // Align to 4-byte boundary with trailing space characters.
+    json.resize(json.size() + (4 - remainder), 0x20);
+  }
+
+  // A binary glTF consists of:
+  // - a 12-byte header
+  // - a JSON chunk (preceded by chunkLength and chunkType)
+  // - a binary buffer chunk (preceded by chunkLength and chunkType)
+  uint32_t totalSize =
+      static_cast<uint32_t>(7 * sizeof(uint32_t) + json.size() + binary.size());
+
+  std::vector<std::byte> result(totalSize);
+  uint32_t* pUint32 = reinterpret_cast<uint32_t*>(result.data());
+  size_t offset = 0;
+  pUint32[offset++] = 0x46546C67; // magic
+  pUint32[offset++] = 2;          // glTF version
+  pUint32[offset++] = totalSize;  // length
+
+  pUint32[offset++] = static_cast<uint32_t>(json.size()); // chunk length
+  pUint32[offset++] = 0x4E4F534A;                         // chunk type (JSON)
+  std::memcpy(pUint32 + offset, json.data(), json.size());
+  offset += json.size() / 4;
+
+  pUint32[offset++] = static_cast<uint32_t>(binary.size()); // chunk length
+  pUint32[offset++] = 0x004E4942;                           // chunk type (BIN)
+  std::memcpy(pUint32 + offset, binary.data(), binary.size());
+
+  return result;
+}
+} // namespace
+
+TEST_CASE("Handles primitive restart during primitive mode conversion") {
+  GltfReader reader;
+
+  SUBCASE("for line loop") {
+    // clang-format off
+    std::vector<uint8_t> indices{
+      255,
+      0, 1, 2, 3, 255,
+      4, 5, 6, 255,
+      7, 8, 9, 10, 255};
+    // clang-format on
+    std::vector<std::byte> file =
+        generateBinaryGltf(CesiumGltf::MeshPrimitive::Mode::LINE_LOOP, indices);
+
+    GltfReaderOptions options;
+    options.primitiveModeOptions.convertLineLoop = true;
+
+    GltfReaderResult result = reader.readGltf(file, options);
+    REQUIRE(result.model);
+    const Model& model = result.model.value();
+
+    REQUIRE_EQ(model.meshes.size(), 1);
+    REQUIRE_EQ(model.meshes[0].primitives.size(), 1);
+
+    const CesiumGltf::MeshPrimitive& primitive = model.meshes[0].primitives[0];
+    CHECK_EQ(primitive.mode, CesiumGltf::MeshPrimitive::Mode::LINES);
+    REQUIRE_EQ(primitive.indices, int32_t(model.accessors.size() - 1));
+
+    CesiumGltf::AccessorView<uint8_t> indicesView(model, primitive.indices);
+    REQUIRE_EQ(indicesView.status(), CesiumGltf::AccessorViewStatus::Valid);
+
+    // clang-format off
+    std::vector<uint8_t> expected{
+      0, 1, 1, 2, 2, 3, 3, 0,
+      4, 5, 5, 6, 6, 4,
+      7, 8, 8, 9, 9, 10, 10, 7};
+    // clang-format on
+    REQUIRE_EQ(indicesView.size(), int64_t(expected.size()));
+    for (int64_t i = 0; i < indicesView.size(); i++) {
+      REQUIRE_EQ(indicesView[i], expected[size_t(i)]);
+    }
+  }
+
+  SUBCASE("for line strip") {
+    // clang-format off
+    std::vector<uint8_t> indices{
+      255,
+      0, 1, 2, 3, 255,
+      4, 5, 6, 255,
+      7, 8, 9, 10, 255};
+    // clang-format on
+    std::vector<std::byte> file = generateBinaryGltf(
+        CesiumGltf::MeshPrimitive::Mode::LINE_STRIP,
+        indices);
+
+    GltfReaderOptions options;
+    options.primitiveModeOptions.convertLineStrip = true;
+
+    GltfReaderResult result = reader.readGltf(file, options);
+    REQUIRE(result.model);
+    const Model& model = result.model.value();
+
+    REQUIRE_EQ(model.meshes.size(), 1);
+    REQUIRE_EQ(model.meshes[0].primitives.size(), 1);
+
+    const CesiumGltf::MeshPrimitive& primitive = model.meshes[0].primitives[0];
+    CHECK_EQ(primitive.mode, CesiumGltf::MeshPrimitive::Mode::LINES);
+    REQUIRE_EQ(primitive.indices, int32_t(model.accessors.size() - 1));
+
+    CesiumGltf::AccessorView<uint8_t> indicesView(model, primitive.indices);
+    REQUIRE_EQ(indicesView.status(), CesiumGltf::AccessorViewStatus::Valid);
+
+    // clang-format off
+    std::vector<uint8_t> expected{
+      0, 1, 1, 2, 2, 3,
+      4, 5, 5, 6,
+      7, 8, 8, 9, 9, 10};
+    // clang-format on
+    REQUIRE_EQ(indicesView.size(), int64_t(expected.size()));
+    for (int64_t i = 0; i < indicesView.size(); i++) {
+      REQUIRE_EQ(indicesView[i], expected[size_t(i)]);
+    }
+  }
+
+  SUBCASE("for triangle strip") {
+    // clang-format off
+    std::vector<uint8_t> indices{
+      255,
+      0, 1, 2, 3, 255,
+      4, 5, 6, 255,
+      7, 8, 9, 10, 255};
+    // clang-format on
+    std::vector<std::byte> file = generateBinaryGltf(
+        CesiumGltf::MeshPrimitive::Mode::TRIANGLE_STRIP,
+        indices);
+
+    GltfReaderOptions options;
+    options.primitiveModeOptions.convertTriangleStrip = true;
+
+    GltfReaderResult result = reader.readGltf(file, options);
+    REQUIRE(result.model);
+    const Model& model = result.model.value();
+
+    REQUIRE_EQ(model.meshes.size(), 1);
+    REQUIRE_EQ(model.meshes[0].primitives.size(), 1);
+
+    const CesiumGltf::MeshPrimitive& primitive = model.meshes[0].primitives[0];
+    CHECK_EQ(primitive.mode, CesiumGltf::MeshPrimitive::Mode::TRIANGLES);
+    REQUIRE_EQ(primitive.indices, int32_t(model.accessors.size() - 1));
+
+    CesiumGltf::AccessorView<uint8_t> indicesView(model, primitive.indices);
+    REQUIRE_EQ(indicesView.status(), CesiumGltf::AccessorViewStatus::Valid);
+
+    // clang-format off
+    std::vector<uint8_t> expected{
+      0, 1, 2, 3, 2, 1,
+      4, 5, 6,
+      7, 8, 9, 10, 9, 8};
+    // clang-format on
+    REQUIRE_EQ(indicesView.size(), int64_t(expected.size()));
+    for (int64_t i = 0; i < indicesView.size(); i++) {
+      REQUIRE_EQ(indicesView[i], expected[size_t(i)]);
+    }
+  }
+
+  SUBCASE("for triangle fan") {
+    // clang-format off
+    std::vector<uint8_t> indices{
+      255,
+      0, 1, 2, 3, 255,
+      4, 5, 6, 255,
+      7, 8, 9, 10, 255};
+    // clang-format on
+    std::vector<std::byte> file = generateBinaryGltf(
+        CesiumGltf::MeshPrimitive::Mode::TRIANGLE_FAN,
+        indices);
+
+    GltfReaderOptions options;
+    options.primitiveModeOptions.convertTriangleFan = true;
+
+    GltfReaderResult result = reader.readGltf(file, options);
+    REQUIRE(result.model);
+    const Model& model = result.model.value();
+
+    REQUIRE_EQ(model.meshes.size(), 1);
+    REQUIRE_EQ(model.meshes[0].primitives.size(), 1);
+
+    const CesiumGltf::MeshPrimitive& primitive = model.meshes[0].primitives[0];
+    CHECK_EQ(primitive.mode, CesiumGltf::MeshPrimitive::Mode::TRIANGLES);
+    REQUIRE_EQ(primitive.indices, int32_t(model.accessors.size() - 1));
+
+    CesiumGltf::AccessorView<uint8_t> indicesView(model, primitive.indices);
+    REQUIRE_EQ(indicesView.status(), CesiumGltf::AccessorViewStatus::Valid);
+
+    // clang-format off
+    std::vector<uint8_t> expected{
+      0, 1, 2, 0, 2, 3,
+      4, 5, 6,
+      7, 8, 9, 7, 9, 10};
     // clang-format on
     REQUIRE_EQ(indicesView.size(), int64_t(expected.size()));
     for (int64_t i = 0; i < indicesView.size(); i++) {